Made the Run button green and renamed cn helper
Restored the green Run button (Primer's primary green) and moved the cn
helper from ui/src/lib/utils.ts to ui/src/cn.ts.
