+++ ui/src/components/alert.tsx
@@ -1,7 +1,7 @@
 import { cva, type VariantProps } from "class-variance-authority";
 import * as React from "react";
 
-import { cn } from "../lib/utils";
+import { cn } from "../cn";
 
 const alertVariants = cva("rounded-md border px-4 py-3 text-sm shadow-sm", {
   variants: {
