+++ ui/src/components/checkbox.tsx
@@ -2,7 +2,7 @@ import { Checkbox as BaseCheckbox } from "@base-ui-components/react/checkbox";
 import { Check } from "lucide-react";
 import * as React from "react";
 
-import { cn } from "../lib/utils";
+import { cn } from "../cn";
 
 const Checkbox = React.forwardRef<React.ElementRef<typeof BaseCheckbox.Root>, React.ComponentPropsWithoutRef<typeof BaseCheckbox.Root>>(
   ({ className, ...props }, ref) => (
