+++ ui/src/components/confirmation-dialog.tsx
@@ -1,7 +1,7 @@
 import { AlertDialog } from "@base-ui-components/react/alert-dialog";
 import * as React from "react";
 
-import { cn } from "../lib/utils";
+import { cn } from "../cn";
 import { buttonVariants } from "./button";
 
 export type ConfirmationGesture = "confirm" | "cancel";
