+++ ui/src/components/dialog.tsx
@@ -2,7 +2,7 @@ import { Dialog as BaseDialog } from "@base-ui-components/react/dialog";
 import { X } from "lucide-react";
 import * as React from "react";
 
-import { cn } from "../lib/utils";
+import { cn } from "../cn";
 import { Button } from "./button";
 
 const widthClass: Record<string, string> = {
