+++ ui/src/components/dropdown-menu.tsx
@@ -1,7 +1,7 @@
 import { Menu } from "@base-ui-components/react/menu";
 import * as React from "react";
 
-import { cn } from "../lib/utils";
+import { cn } from "../cn";
 
 const DropdownMenu = Menu.Root;
 const DropdownMenuGroup = Menu.Group;
