+++ ui/src/components/icon-button.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 
-import { cn } from "../lib/utils";
+import { cn } from "../cn";
 import type { Icon } from "./icons";
 import { buttonVariants } from "./button";
 import { SimpleTooltip } from "./tooltip";
